Widen sendResponse()/sendResponseUnencrypted() length to uint16_t (#63)
sendResponse() assigns the encrypted length into its `len` parameter:

    uint16_t encrypted_len = 0;
    ... encryptResponse(response, len, encrypted_response, &encrypted_len, ...);
    len = encrypted_len;

but `len` was uint8_t, so an encrypted response (up to 2+16+254+12 = 284 bytes)
truncates to len & 0xFF. No plaintext response currently exceeds ~100 bytes so
it is not yet triggered, but it is a latent silent-corruption bug and the
encrypted_response[600] buffer implies large responses were intended.

Widen `len` to uint16_t on both functions and every forward declaration
(communication.h/.cpp, main.h, encryption.cpp, buzzer_control.cpp,
device_control.cpp) so callers still resolve the same symbol.

+++ src/buzzer_control.cpp
@@ -4,7 +4,7 @@
 #include <string.h>
 
 extern struct GlobalConfig globalConfig;
-void sendResponse(uint8_t* response, uint8_t len);
+void sendResponse(uint8_t* response, uint16_t len);
 
 static_assert(sizeof(PassiveBuzzerConfig) == 32, "PassiveBuzzerConfig must be 32 bytes");
