+++ src/communication.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-void sendResponseUnencrypted(uint8_t* response, uint8_t len);
-void sendResponse(uint8_t* response, uint8_t len);
+void sendResponseUnencrypted(uint8_t* response, uint16_t len);
+void sendResponse(uint8_t* response, uint16_t len);
 uint16_t calculateCRC16CCITT(uint8_t* data, uint32_t len);
 uint8_t getFirmwareMajor();
 uint8_t getFirmwareMinor();
